// src/oncobridge-web/src/app/api/models/patient-timeline-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TimelineGroupResponse } from '../models/timeline-group-response';
import { TimelineProjectionPolicyResponse } from '../models/timeline-projection-policy-response';
import { UnsequencedTimelineEventResponse } from '../models/unsequenced-timeline-event-response';
export interface PatientTimelineResponse {
  groups: Array<TimelineGroupResponse>;
  patientId: string;
  projectionPolicy: TimelineProjectionPolicyResponse;
  unsequencedEvents: Array<UnsequencedTimelineEventResponse>;
}